requirements.txt
# Web Framework
fastapi==0.115.5
uvicorn==0.32.1
jinja2==3.1.4
python-multipart==0.0.20

# HTTP Client
requests==2.32.3

# Browser Automation
selenium==4.27.1

# TOTP Authentication
pyotp==2.9.0

# Environment Variables
python-dotenv==1.0.0

# Image Processing
pillow==11.0.0
opencv-python==4.10.0.84
numpy>=1.21.0,<2.0.0
